PSC: copy Harbor token to volume mount
This change copies Harbor's tokens.properties file to
/data/harbor/psc. This is needed to support Harbor's standalone
use-case where the user may not have access to /etc while
attempting to mount the token file's path as a volume in the
Harbor container.

To avoid a systemd dependency cycle with the harbor_startup.path
unit, this change creates a marker path after the token file has
been copied and uses that path to activate harbor_startup.service.