Add support for upgrading OVA 1.2.x to 1.2.y (#871)
* Add support for upgrading OVA 1.2.x to 1.2.y

This commit modifies and renames the upgrade_1.1_to_1.2.sh script
to upgrade.sh. The script now supports upgrading from 1.2.x to
1.2.y as well.

The script now detects the presence of Admiral's PSC config file
(/data/admiral/configs/psc-config.properties) to detect the version
of the older OVA. If this file exists on the data disk, the OVA must
be of version 1.2.x. Then, data migration is skipped but Harbor's
address is updated in Admiral configuration.

* Add confirmation prompt to upgrade script

This change adds a confirmation prompt after detecting the old OVA
version before continuing with upgrade.