Add tutorial for VCH Management API (#1004)
An API for VCH Management is under development and will be included in
1.3.0. While the current functionality targets a very specific use-
case: enabling the implementation of a VCH creation wizard in the H5
client plugin, the API will be publicly exposed and accessible.

This tutorial attempts to explain that this pre-1.0 API is not ready
for general consumption, but that it is available for experimentation
and feedback is welcome. The earlier we can get feedback from future
users of the API, the better able we will be to accommodate it.