Automate VIC UI Plugin Install (#1758 #1766 #1785 #1799)
Introduces caching of the vic-ui plugin artifacts as a component
of the ova. Updates the fileserver provisioner to package vic-ui
with the correct version numbers.

Creates the Fileserver Thumbprint API to return the
thumbprint of a given target ip address or fqdn.
Also refactors the fileserver main process into
main.go, and creates the routes/ and tasks/
sub-packages.

Moves vic-ui plugin dependencies to fileserver. This includes:
 - vic/lib/install/ova
 - vic/lib/install/plugin
- vic/tagvm/
Additionally, refactors ovatools/vic-ui/ui/ui.go to plugin.go that
doesn't include any CLI flags.

Automates the plugin install process and improves
logging in the fileserver:
 - Add trace.Op logging to fileserver tasks
 - Require vcenter thumprint on psc registration
 - Install VIC UI plugins in the initialization process
 - Attaches a managed tag to the Appliance VM
 - Performs a plugin install during the appliance upgrade
   process using the /plugin API.

Fixes #1433, #637, #1720, #1702 and #1789.