Crossport OVA-Util, VM-Util, and Util from master (#1842)
There have been a large number of tangled commits to master that
modify various sets of robot resources and tests. The right thing to
do would be to tease those apart and identify the pieces of each which
are needed to fix deployment of the OVA on 1.4.1. Doing so has proved
difficult, so this change crossports the current state of three key
files.

This will very likely break nightly tests on the release branch in
a variety of ways. This is acceptable because we don't currently run
nightly tests on the release branch.