Remove erroneous backticks in linker flags
When replacing the tag command with an environment variable, the
surrounding backticks were not removed.

This was not caught during testing because the script did not return
a non-zero exit code, even though an error had occurred. Add a set
line to prevent similar errors from going undetected in the future.