Revise upgrade script to treat v1.4.0 like v1.3.x
Because v1.2.1 had a different disk topology than v1.3.0 and later, it
requires a slightly different upgrade process. This was handled by
explicitly checking whether the "source" version is v1.3.0 or v1.3.1
and moving the new disks only in those cases.

We expect the disk topology to remain consistent until an intentional
change is made. Reverse the conditional logic to disable the v1.3.0+
logic only when the version is v1.2.1. (Upgrade from earlier versions
is unsupported, and an error is returned if attempted.)

This causes upgrade from v1.4.0 to behave like upgrade from v1.3.x and
ensures that support for upgrade from future v1.4.x patches can easily
be added without needing to update several places in this file.