Prevent password masking from becoming password (#1873)
There is a path, likely retry or recoverying but at this time unconfirmed,
by which the password masking can be set as the password. This expressly
checks for the value having been masked and refuses to change the password
to that value.

This is a very focused change to address specifically the observed
behaviour as the implications of a revert or moving the code to other
units are not understood at this time.
Issue was introduced in 3c5cd51a
(cherry picked from commit 51844757c1dd73af5c8510d0f3e75839ce25340f)