Output reminder for next step of release process (#1860)
As described in installer/docs/BUILD.md, our release process involves
several steps: a tag build, a staging build, and a release build.

Fully connecting these steps with automation is undesirable; we don't
want to stage all tag builds and we want to manually inspect staged
builds before deciding to release them.

However, requiring users to manually assemble the appropriate command
for the next step of the process is error prone. Instead, we output a
script containing the next command for them to copy and run.