Print file size and checksums before publishing (#1867) (#1875)
As documented in installer/docs/RELEASE.md, our process is to include
filesize and checksum information when publishing a release on GitHub.

Output that information in a structured way during the build process
to ensure that accurate information can be easily posted.

Include this output for all published builds to allow integrity to be
verified in all cases.

(cherry picked from commit 4459076366ff757c958a4fd71ba459fdab9f9509)