Work around inexplicable pipe error during build (#1888)
For some reason, `head -n1` is resulting in exit code 141 despite being
protected by `trap '' PIPE`. This same pattern is used successfully in
the preceding logic. It is not clear why we are seeing this issue now,
or why trap is not behaving as expected.

This issue cannot be reproduced locally.

Disabling pipefail for this line is acceptable because the build does
not even use the results of this line on the release branch.

Note: This change should be reverted once the cause is understood.