Remove "-cert" from the image tag used in the create example (#1915)
* Added missing "-it" option in the example

The dch-photon needs to be deployed with "-it" option. Without out this option the container will not respond to the docker-client when trying to use the dch.

* Corrected dch-photon image tag

The default tag for the dch-photon image is 1.13. The tag 1.13-cert seems to be coming from the alternative method of building the certificates directly into the image which is not a prerequisite for following the manual certificate procedure.

* Removed unnecessary option during create example