Fix construction of OVA URL in scenario test
Download link is being synthesized using product branch, which may be left
underfined. In this case we have // in the URL. Google is sensitive to
this double slash so we must sanitize the path before use.

This also fixes the error handling for total failure to download after the
retries.