Document dependency consumption per build type (#1950)
The way we determine the build of each dependency to consume depends
on both on the way the dependency's artifacts are published and the
type of build being performed. Document the current behavior in
preparation for proposing changes.

Additionally, document the variety of locations where the OVA may be
published depending on the type and results of the build.

(cherry picked from commit 6f68879c1db40f79d153ad7d3cbd50144f86814b)