Integrate pre-tag components in release builds (#1961)
Consuming only tagged components in non-dev tag builds ensures that
our release candidates and releases include only the expected, stable,
version of each component.

Consuming those same tagged components in push and pull_request builds
is beneficial because it maintains consistency between builds leading
up to a tag build and the tag build itself.

However, this has a significant disadvantage: it means that components
being developed on release branches concurrently are not continuously
integrated; the first time integration occurs is when RC1 is tagged.

To allow testing of OVA builds sooner, we can consume new versions of
components more aggressively for pull_request builds, push builds, and
dev tag builds.

For Admiral, we will consume the same vic_dev tag consumed on master.

For the VIC engine .tar.gz and vic-machine-server container image, we
consume the latest build from the corresponding release branch. Note
that we assume that the version number of the OVA is the same as the
version number of VIC. To produce a release where this assumption is
not not valid (e.g., a patch release of VIC product where the version
of VIC itself is not changed), the script could be modified directly
on the release branch to hard-code the appropriate VIC version.