Update VCH Management API tutorial for 1.4 (#1991)
A tutorial on the use of the VCH Management API was written while the
API was under development for 1.3.0 and includes a variety of stale
information as a result. The tutorial is still a useful starting point
for users interested in providing feedback on the API. Update
instructions and documentation links to be relevant for 1.4.x.