Quote potentially empty param to nimbus keyword (#2017)
If omitting the ESX and vCenter version parameters to the Nimbus Testbed
Deploy keyword we get arguments to the underlying script being shifted
and causing confusing output/failures.

This quotes the provided strings to ensure that even if empty we keep
the argument positions when calling the underlying scripts. We will still
fail but the error output is at least sane in that condition.