Add internal no proxy and keep custom.config
custom.config is for user 's customized configurations and should
not be overwritten. Use it as the base context of config.properties
and append auto generated configurations.

localhost, 127.0.0.1, ui, {HOSTNAME} is important for the internal
connections on the VIC Appliance between the Admiral service and
the Harbor registry. Include them in no proxy list whenever proxy
is configured.

When proxy is not set or removed in the vApp Options, set them as
null clearly.